Updated package header URL
I noticed that the URL field of the package header does not refer to a web address.  I can't find any source documenting it's "official" use, but Marmalade refers to it (https://marmalade-repo.org/) and it is needed by `paradox-menu-visit-homepage` to be a proper URL.

Thanks for publishing the package!  :+1: